Fazendo otimizações no projeto
