+++ src/App.css
@@ -83,5 +83,4 @@
     --pf-global--BackgroundColor--dark-400: #3B8952;
     --pf-global--BackgroundColor--dark-transparent-100: rgba(61, 19, 0, 0.62); 
     --pf-global--BackgroundColor--dark-transparent-200: rgba(61, 19, 0, 0.62);
-
 }
