+++ src/features/Estacoes/context.js
@@ -1,12 +1,26 @@
-import React, { createContext, useState } from 'react'
-import api from '../../services/api'
-import { useQuery } from 'react-query'
+import React, { createContext,useEffect, useState } from 'react';
+import api from '../../services/api';
+import { useQuery } from 'react-query';
+import { useParams } from "react-router-dom";
+import placeholderData from "./placeholderData";
 
 const context = createContext(true)
+const arrPlaceHolder = [placeholderData];
 
 export const Provider = ({ children }) => {
   const [estacaoId] = useState('A301')
-  const [estacao] = useState()
+  const {atributoId, initialDateId, finalDateId, codEstacaoId } = useParams();
+  const [estacao, setEstacao] = useState("A301 - RECIFE");
+  const [initialDate, setInitialDate] = useState("01-10-2019");
+  const [finalDate, setFinalDate] = useState("31-10-2019");
+  const [atributo, setAtributo] = useState("CHUVA");
+  const [initialDateFormat, setInitialDateFormat] = useState(initialDateId? initialDateId : "2019-10-01");
+  const [finalDateFormat, setFinalDateFormat] = useState(finalDateId? finalDateId: "2019-10-31");
+  const [codEstacao, setCodEstacao] = useState(codEstacaoId? codEstacaoId: "A301");
+  const [atributoFinal, setAtributoFinal] = useState(atributoId? atributoId: "CHUVA");
+  const [title, setTitle] = useState(codEstacaoId? codEstacaoId:"A301 - RECIFE");
+  const [validador, setValidador] = useState(true);
+  const [num, setNum] = useState(0);
 
   const getEstacaoById = eid => estacoes.find(e => e.CD_ESTACAO === eid)
 
@@ -23,6 +37,34 @@ export const Provider = ({ children }) => {
     }
   )
 
+  const { data: dataEstacao, isFetching } = useQuery(
+    ["dataPlot", initialDateFormat, finalDateFormat, codEstacao],
+    () =>
+      api
+        .get(
+          `estacao/diaria/${initialDateFormat}/${finalDateFormat}/${codEstacao}`
+        )
+        .then((r) => r.data),
+    {
+      onError: (error) => {
+        console.log("erro no fetch dos dados: " + error);
+      },
+      onSuccess: () => {
+        console.log("fetch sucess");
+      },
+      refetchOnWindowFocus: false,
+      retry: false,
+      staleTime: 1000 * 60 * 5,
+      placeholderData: arrPlaceHolder.at(-1),
+    }
+  );
+
+  useEffect(() => {
+    if (!dataEstacao) return null;
+    
+    arrPlaceHolder.push(dataEstacao);
+  });
+
   const estacoesIsLoading = estacoesQuery.isLoading
   const estacoes = estacoesQuery.data
 
@@ -31,7 +73,30 @@ export const Provider = ({ children }) => {
     estacoes,
     estacoesIsLoading,
     getEstacaoById,
-    estacaoId
+    estacaoId,
+    setEstacao,
+    initialDate,
+    setInitialDate,
+    finalDate,
+    setFinalDate,
+    dataEstacao,
+    initialDateFormat,
+    setInitialDateFormat,
+    finalDateFormat,
+    setFinalDateFormat,
+    codEstacao,
+    setCodEstacao,
+    atributo,
+    setAtributo,
+    atributoFinal,
+    setAtributoFinal,
+    validador,
+    setValidador,
+    isFetching,
+    title,
+    setTitle,
+    num,
+    setNum,
   }
   return <context.Provider value={provides}>{children}</context.Provider>
 }
