+++ src/features/Estacoes/index.js
@@ -1,18 +1,51 @@
 import Layout from "../shared/Layout";
 import { useContext } from "react";
 import context, { Provider } from "./context";
-import { FilterProvider } from "./filterContext";
 import FilterForm from "./FilterForm";
 import ChartLine from "./ChartLine";
-import { ReactQueryDevtools } from "react-query/devtools";
 import PlotMapbox from "./PlotMapbox";
-import { Grid, GridItem } from "@patternfly/react-core";
+import { Grid, GridItem, Skeleton, Spinner } from "@patternfly/react-core";
 
-const Estacoes = () => {
+const PageEstacoes = () => {
   const { estacoesIsLoading } = useContext(context);
+  // const estacoesIsLoading = true;
 
   if (estacoesIsLoading) {
-    return <Layout>Teste 2</Layout>;
+    return (
+      <Layout>
+        <Skeleton
+          fontSize="4xl"
+          screenreaderText="Carregando..."
+          className="pf-c-skeleton"
+        />
+        <br />
+        <div style={{ display: "flex" }}>
+          <div
+            style={{
+              height: "70vh",
+              width: "50%",
+              display: "flex",
+              alignItems: "flex-end",
+              justifyContent: "space-between",
+            }}
+          >
+            <Skeleton height="25%" width="15%" />
+            <Skeleton height="33%" width="15%" />
+            <Skeleton height="50%" width="15%" />
+            <Skeleton height="66%" width="15%" />
+            <Skeleton height="75%" width="15%" />
+            <Skeleton height="100%" width="15%" />
+          </div>
+          <div className="pf-u-m-auto">
+            <Spinner
+              isSVG
+              diameter="200px"
+              aria-label="Contents of the custom size example"
+            />
+          </div>
+        </div>
+      </Layout>
+    );
   }
 
   return (
@@ -25,13 +58,10 @@ const Estacoes = () => {
   );
 };
 
-const index = () => (
+const Estacoes = () => (
   <Provider>
-    <FilterProvider>
-      <Estacoes />
-      <ReactQueryDevtools />
-    </FilterProvider>
+    <PageEstacoes />
   </Provider>
 );
 
-export default index;
+export default Estacoes;
