+++ src/features/shared/utils/dateFormat.js
@@ -1,0 +1,10 @@
+const dateFormat = (date) =>
+  date
+    .toLocaleDateString("pt-BR", {
+      year: "numeric",
+      month: "2-digit",
+      day: "2-digit",
+    })
+    .replace(/\//g, "-");
+
+export default dateFormat;
